removed multiplexer so you dont have to send route down, this isnt very thread safe...sigh
