+++ src/Ocelot/DownstreamRouteFinder/Middleware/DownstreamRouteFinderMiddleware.cs
@@ -1,4 +1,5 @@
 using System.Threading.Tasks;
+using Ocelot.Configuration;
 using Ocelot.Configuration.Provider;
 using Ocelot.DownstreamRouteFinder.Finder;
 using Ocelot.Infrastructure.Extensions;
@@ -58,9 +59,48 @@ namespace Ocelot.DownstreamRouteFinder.Middleware
             //todo - put this back in
             // _logger.LogDebug("downstream template is {downstreamRoute.Data.ReRoute.DownstreamPath}", downstreamRoute.Data.ReRoute.DownstreamReRoute.DownstreamPathTemplate);
 
-            context.DownstreamRoute = downstreamRoute.Data;
+            context.TemplatePlaceholderNameAndValues = downstreamRoute.Data.TemplatePlaceholderNameAndValues;
 
+            await Multiplex(context, downstreamRoute.Data.ReRoute);
+        }
+
+        private async Task Multiplex(DownstreamContext context, ReRoute reRoute)
+        {
+            var tasks = new Task<DownstreamContext>[reRoute.DownstreamReRoute.Count];
+            for (int i = 0; i < reRoute.DownstreamReRoute.Count; i++)
+            {
+                var downstreamContext = new DownstreamContext(context.HttpContext)
+                {
+                    TemplatePlaceholderNameAndValues = context.TemplatePlaceholderNameAndValues,
+                    ServiceProviderConfiguration = context.ServiceProviderConfiguration,
+                    DownstreamReRoute = reRoute.DownstreamReRoute[i],
+                    //todo do we want these set here
+                    RequestId = context.RequestId,
+                    PreviousRequestId = context.PreviousRequestId,
+                };
+
+                tasks[i] = Fire(downstreamContext);
+            }
+
+            await Task.WhenAll(tasks);
+
+            //now cast the complete tasks to whatever they need to be
+            //store them and let the response middleware handle them..
+
+            var finished = tasks[0].Result;
+
+            context.Errors = finished.Errors;
+            context.DownstreamRequest = finished.DownstreamRequest;
+            context.DownstreamResponse = finished.DownstreamResponse;
+            context.RequestId = finished.RequestId;
+            context.PreviousRequestId = finished.RequestId;
+            
+        }
+
+        private async Task<DownstreamContext> Fire(DownstreamContext context)
+        {
             await _next.Invoke(context);
+            return context;
         }
     }
 }
