+++ src/Ocelot/Middleware/Pipeline/OcelotPipelineExtensions.cs
@@ -36,8 +36,6 @@ namespace Ocelot.Middleware.Pipeline
             // Then we get the downstream route information
             builder.UseDownstreamRouteFinderMiddleware();
 
-            builder.UseMultiplexerMiddleware();
-
             // Now we have the ds route we can transform headers and stuff?
             builder.UseHttpHeadersTransformationMiddleware();
