+++ test/Ocelot.UnitTests/Authorization/AuthorisationMiddlewareTests.cs
@@ -62,7 +62,7 @@ namespace Ocelot.UnitTests.Authorization
 
         private void GivenTheDownStreamRouteIs(DownstreamRoute downstreamRoute)
         {
-            _downstreamContext.DownstreamRoute = downstreamRoute;
+            _downstreamContext.TemplatePlaceholderNameAndValues = downstreamRoute.TemplatePlaceholderNameAndValues;
             _downstreamContext.DownstreamReRoute = downstreamRoute.ReRoute.DownstreamReRoute[0];
         }
