+++ test/Ocelot.UnitTests/Claims/ClaimsBuilderMiddlewareTests.cs
@@ -68,7 +68,7 @@ namespace Ocelot.UnitTests.Claims
 
         private void GivenTheDownStreamRouteIs(DownstreamRoute downstreamRoute)
         {
-            _downstreamContext.DownstreamRoute = downstreamRoute;
+            _downstreamContext.TemplatePlaceholderNameAndValues = downstreamRoute.TemplatePlaceholderNameAndValues;
             _downstreamContext.DownstreamReRoute = downstreamRoute.ReRoute.DownstreamReRoute[0];
         }
