+++ test/Ocelot.UnitTests/QueryStrings/QueryStringBuilderMiddlewareTests.cs
@@ -90,7 +90,7 @@ namespace Ocelot.UnitTests.QueryStrings
 
         private void GivenTheDownStreamRouteIs(DownstreamRoute downstreamRoute)
         {
-            _downstreamContext.DownstreamRoute = downstreamRoute;
+            _downstreamContext.TemplatePlaceholderNameAndValues = downstreamRoute.TemplatePlaceholderNameAndValues;
             _downstreamContext.DownstreamReRoute = downstreamRoute.ReRoute.DownstreamReRoute[0];
         }
     }
