+++ test/Ocelot.UnitTests/RateLimit/ClientRateLimitMiddlewareTests.cs
@@ -82,7 +82,7 @@ namespace Ocelot.UnitTests.RateLimit
 
         private void GivenTheDownStreamRouteIs(DownstreamRoute downstreamRoute)
         {
-            _downstreamContext.DownstreamRoute = downstreamRoute;
+            _downstreamContext.TemplatePlaceholderNameAndValues = downstreamRoute.TemplatePlaceholderNameAndValues;
             _downstreamContext.DownstreamReRoute = downstreamRoute.ReRoute.DownstreamReRoute[0];
         }
