+++ test/Ocelot.UnitTests/RequestId/ReRouteRequestIdMiddlewareTests.cs
@@ -160,7 +160,7 @@ namespace Ocelot.UnitTests.RequestId
 
         private void GivenTheDownStreamRouteIs(DownstreamRoute downstreamRoute)
         {
-            _downstreamContext.DownstreamRoute = downstreamRoute;
+            _downstreamContext.TemplatePlaceholderNameAndValues = downstreamRoute.TemplatePlaceholderNameAndValues;
             _downstreamContext.DownstreamReRoute = downstreamRoute.ReRoute.DownstreamReRoute[0];
         }
